server: updated functionality change password, removed unnecessary function addUser

+++ server/src/controllers/student.controller.ts
@@ -1,23 +1,14 @@
 import { Request, Response } from 'express';
 import UserService from '../services/student.service';
+import ApiError from '../utils/api-error';
 
 class StudentController {
-  async addUser(req: Request, res: Response) {
-    try {
-      const response = await UserService.addUser();
-      return res.status(200).json({
-        data: response,
-      });
-    } catch (e: any) {
-      return res.status(400).json({
-        data: e.message,
-      });
-    }
-  }
-
   async removeUser(req: Request, res: Response) {
     try {
       const { userId } = req.body;
+      if (!userId) {
+        throw ApiError.BadRequestError('You have to add send userId');
+      }
       const response = await UserService.removeUser(userId);
       return res.status(200).json({
         data: response,
